Update understanding-protection-level.md
Hi Team,

I found a bit of unnatural sentences in Japanese language, and I would like to fix it. This proposal is supposed to be addressed as a Linguistic Suggestion from my point of view. It would be appreciated if you could take a look at this pull-request. Please let me know if you have any concerns.

This request is in regard to:

- https://docs.microsoft.com/ja-jp/dotnet/framework/wcf/understanding-protection-level
- https://docs.microsoft.com/en-us/dotnet/framework/wcf/understanding-protection-level

Best regards,